+++ src/main/java/it/objectmethod/e/commerce/repository/ArticoloRepository.java
@@ -9,7 +9,8 @@ import org.springframework.stereotype.Repository;
 import it.objectmethod.e.commerce.entity.Articolo;
 
 @Repository
-public interface ArticoloRepository extends JpaRepository<Articolo, Integer> {
+public interface ArticoloRepository extends JpaRepository<Articolo, Long> {
 	@Query("select c from Articolo c where (''=?1 or c.nomeArticolo=?1) and (''=?2 or c.codiceArticolo = ?2)")
 	public List<Articolo> trovaArticoli(String name, String codiceArticolo);
+
 }
